Bump to -1

+++ build.sh
@@ -11,29 +11,34 @@ echo "Branch in jonathanusername/flow is: $BRANCH"
 echo "Extra suffix to eventual version is: $EXTRA"
 echo "Full version to include in package.json will be: $VERSION"
 
-SRC_DIR='flow-src'
-if [ -d $SRC_DIR ]; then
-    rm -rf $SRC_DIR
-fi;
 
-git clone git@github.com:JonathanUsername/flow.git $SRC_DIR;
+if [[ -z "$NO_MAC" ]]; then
+    SRC_DIR='flow-src'
+    if [ -d $SRC_DIR ]; then
+        rm -rf $SRC_DIR
+    fi;
 
-pushd $SRC_DIR;
+    git clone git@github.com:JonathanUsername/flow.git $SRC_DIR;
 
-# I'm giving up on old way of applying patches, easier to adjust in jonathanusername/flow, make a new branch with the right version and then just pull the new version here, but for posterity, I used to do it like this:  && git am -3 < ../changes.patch && make || exit 1
+    pushd $SRC_DIR;
 
-git fetch && git checkout "$BRANCH" && make || exit 1 
+    # I'm giving up on old way of applying patches, easier to adjust in jonathanusername/flow, make a new branch with the right version and then just pull the new version here, but for posterity, I used to do it like this:  && git am -3 < ../changes.patch && make || exit 1
 
-popd;
+    git fetch && git checkout "$BRANCH" && make || exit 1 
 
-mv flow-src/bin/flow dist/mac/flow
-echo 'Mac binaries installed'
+    popd;
 
-echo 'Building linux binaries in docker. If this fails, try building with --no-cache and re-running'
-docker build --no-cache -t flow-bin .
-docker run -it  -v `pwd`/dist/linux:/root/flow/bin -e "FLOW_VERSION=$BRANCH" flow-bin:latest '/root/linux-build.sh'
-echo 'Linux binaries installed' 
+    mv flow-src/bin/flow dist/mac/flow
+    echo 'Mac binaries installed'
+fi
 
+
+if [[ -z "$NO_LINUX" ]]; then
+    echo 'Building linux binaries in docker. If this fails, try building with --no-cache and re-running (it will be a LOT slower)'
+    docker build -t flow-bin .
+    docker run -it  -v `pwd`/dist/linux:/root/flow/bin -e "FLOW_VERSION=$BRANCH" flow-bin:latest 
+    echo 'Linux binaries installed' 
+fi
 
 rewrite_package_version() {
   NEWPACK=$(cat package.json | jq "to_entries | map(if .key == \"version\" then . + {\"value\": \"$VERSION\"} else . end) | from_entries")
